Add custom query support in redis plugin.
With this patch the redis.c plugin gains the capability to query the redis
server for a value (like HLEN, ZLEN and so on), and return the value (it
must be a single value, multiple values are not allowed) to be printed.

This patch require previous libhiredis support for redis.c plugin.